# openapi/bundled/public.yaml
openapi: 3.0.3
info:
  title: Mocó Public API
  version: 0.2.0
  description: |
    Tenant-isolated, path-based secret management for interactive and
    automated clients. The contract makes no network-topology assumptions:
    clients can run inside or outside the Mocó Kubernetes cluster.
  contact:
    name: AraiHu
    url: https://github.com/araihu/moco
servers:
  - url: /
    description: Current deployment origin. Production deployments use TLS.
security:
  - bearerAuth: []
...
paths:
...
  /api/v1/watch:
    parameters:
      - $ref: '#/components/parameters/request-id'
    get:
      operationId: watchChanges
      summary: Wait for persisted resource changes
      description: |
        Returns a durable, process-wide resource version. With no
        `resourceVersion`, the endpoint returns the current checkpoint immediately.
        With a checkpoint, it returns immediately when the checkpoint is stale or
        waits up to `timeoutSeconds` for a newer revision. A true `changed` result
        is a signal to resync the tenant, vault, and secret collections; this
        endpoint does not expose payloads, ordering, or tombstones. The checkpoint
        is process-wide, so this operation requires an explicit cluster-wide watch
        permission (`domain: "*"`); a tenant-scoped policy cannot authorize it.
      tags: [Service]
      parameters:
        - $ref: '#/components/parameters/resource-version'
        - $ref: '#/components/parameters/watch-timeout'
      responses:
        '200':
          description: Current checkpoint and whether it advanced.
          headers:
            Cache-Control:
              $ref: '#/components/headers/cache-control'
            X-Request-ID:
              $ref: '#/components/headers/request-id'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/watch-result'
        '400':
          $ref: '#/components/responses/bad-request'
        '401':
          $ref: '#/components/responses/unauthorized'
        '403':
          $ref: '#/components/responses/forbidden'
        '429':
          $ref: '#/components/responses/too-many-requests'
        '500':
          $ref: '#/components/responses/internal-error'
        '503':
          $ref: '#/components/responses/service-unavailable'
  /api/v1/tenants:
    parameters:
      - $ref: '#/components/parameters/request-id'
    get:
      operationId: listTenants
      summary: List tenants
      description: |
        Lists tenants in the caller's collection authorization scope. The first
        request creates a sequence-bounded snapshot; follow `nextCursor` until
        `hasMore` is false. If a cursor expires, restart without one. Name and
        externalId filters are exact and may be combined. For deterministic
        adoption, controllers should use an exact immutable `externalId` filter
        before creating a tenant.
      tags: [Tenants]
      parameters:
        - $ref: '#/components/parameters/limit'
        - $ref: '#/components/parameters/cursor'
        - $ref: '#/components/parameters/name-filter'
        - $ref: '#/components/parameters/external-id-filter'
      responses:
        '200':
          description: Page of tenants ordered by immutable insertion sequence.
          headers:
            X-Request-ID:
              $ref: '#/components/headers/request-id'
          content:
            application/json:
              example:
                items:
                  - id: 11111111-1111-4111-8111-111111111111
                    name: production
                    labels: {}
                    revision: 1
                    createdAt: '2026-08-25T12:00:00Z'
                    updatedAt: '2026-08-25T12:00:00Z'
                page:
                  hasMore: false
                  nextCursor: null
              schema:
                $ref: '#/components/schemas/tenant-list'
        '400':
          $ref: '#/components/responses/bad-request'
        '401':
          $ref: '#/components/responses/unauthorized'
        '403':
          $ref: '#/components/responses/forbidden'
        '410':
          $ref: '#/components/responses/cursor-expired'
        '429':
          $ref: '#/components/responses/too-many-requests'
        '500':
          $ref: '#/components/responses/internal-error'
        '503':
          $ref: '#/components/responses/service-unavailable'
    post:
      operationId: createTenant
      summary: Create a tenant
      description: |
        Creates a tenant. Name and externalId are independently unique. Automation
        should supply an externalId and Idempotency-Key. A uniqueness conflict
        returns the existing resourceId when safe, allowing an atomic adoption
        handshake. Repeating an identical request with the same Idempotency-Key
        returns the original result; reusing that key with another body returns 409.
      tags: [Tenants]
      parameters:
        - $ref: '#/components/parameters/idempotency-key'
      requestBody:
        description: Tenant creation fields.
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/tenant-create'
      responses:
        '201':
          description: Tenant created, or prior identical keyed result replayed.
          headers:
            ETag:
              $ref: '#/components/headers/etag'
            Location:
              $ref: '#/components/headers/location'
            X-Request-ID:
              $ref: '#/components/headers/request-id'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/tenant'
        '400':
          $ref: '#/components/responses/bad-request'
        '401':
          $ref: '#/components/responses/unauthorized'
        '403':
          $ref: '#/components/responses/forbidden'
        '409':
          $ref: '#/components/responses/conflict'
        '429':
          $ref: '#/components/responses/too-many-requests'
        '500':
          $ref: '#/components/responses/internal-error'
        '503':
          $ref: '#/components/responses/service-unavailable'
  /api/v1/tenants/{tenantId}:
    parameters:
      - $ref: '#/components/parameters/tenant-id'
      - $ref: '#/components/parameters/request-id'
    get:
      operationId: getTenant
      summary: Get a tenant
      description: Returns one tenant by its stable ID.
      tags: [Tenants]
      parameters:
        - $ref: '#/components/parameters/if-none-match'
      responses:
        '200':
          description: Tenant representation.
          headers:
            ETag:
              $ref: '#/components/headers/etag'
            X-Request-ID:
              $ref: '#/components/headers/request-id'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/tenant'
        '304':
          $ref: '#/components/responses/not-modified'
        '400':
          $ref: '#/components/responses/bad-request'
        '401':
          $ref: '#/components/responses/unauthorized'
        '403':
          $ref: '#/components/responses/forbidden'
        '404':
          $ref: '#/components/responses/not-found'
        '429':
          $ref: '#/components/responses/too-many-requests'
        '500':
          $ref: '#/components/responses/internal-error'
        '503':
          $ref: '#/components/responses/service-unavailable'
    put:
      operationId: updateTenant
      summary: Update a tenant
      description: |
        Replaces all mutable fields. externalId and server fields are immutable.
        If-Match is optional; omit it only for an intentionally unconditional
        mutation. Retrying clients should read the representation and use its ETag.
      tags: [Tenants]
      parameters:
        - $ref: '#/components/parameters/if-match'
      requestBody:
        description: Complete mutable tenant state. Use null to clear description.
        required: true
        content:
          application/json:
            example:
              name: platform
              description: Platform-owned tenant
              labels:
                app.kubernetes.io/managed-by: moco-operator
            schema:
              $ref: '#/components/schemas/tenant-update'
      responses:
        '200':
          description: Updated tenant representation.
          headers:
            ETag:
              $ref: '#/components/headers/etag'
            X-Request-ID:
              $ref: '#/components/headers/request-id'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/tenant'
        '400':
          $ref: '#/components/responses/bad-request'
        '401':
          $ref: '#/components/responses/unauthorized'
        '403':
          $ref: '#/components/responses/forbidden'
        '404':
          $ref: '#/components/responses/not-found'
        '409':
          $ref: '#/components/responses/conflict'
        '412':
          $ref: '#/components/responses/precondition-failed'
        '429':
          $ref: '#/components/responses/too-many-requests'
        '500':
          $ref: '#/components/responses/internal-error'
        '503':
          $ref: '#/components/responses/service-unavailable'
    delete:
      operationId: deleteTenant
      summary: Delete a tenant
      description: |
        Deletes an empty tenant. With cascade=true, also deletes all child vaults
        and secrets. A missing tenant returns 404; reconcilers may treat it as
        success. If-Match is optional, but recommended for retries that must not
        delete a newer revision.
      tags: [Tenants]
      parameters:
        - $ref: '#/components/parameters/if-match'
        - $ref: '#/components/parameters/cascade'
      responses:
        '204':
          description: Tenant deleted.
          headers:
            X-Request-ID:
              $ref: '#/components/headers/request-id'
        '400':
          $ref: '#/components/responses/bad-request'
        '401':
          $ref: '#/components/responses/unauthorized'
        '403':
          $ref: '#/components/responses/forbidden'
        '404':
          $ref: '#/components/responses/not-found'
        '409':
          $ref: '#/components/responses/conflict'
        '412':
          $ref: '#/components/responses/precondition-failed'
        '429':
          $ref: '#/components/responses/too-many-requests'
        '500':
          $ref: '#/components/responses/internal-error'
        '503':
          $ref: '#/components/responses/service-unavailable'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: |
        Deployment-issued bearer credential sent in the Authorization header.
        Tokens may be opaque; clients must not assume JWT structure.
  schemas:
...
    watch-result:
      title: WatchResult
      type: object
      description: Durable resource-version checkpoint returned by a watch poll.
      additionalProperties: false
      required:
        - resourceVersion
        - changed
      properties:
        resourceVersion:
          type: string
          pattern: '^rv-[0-9]+$'
          minLength: 3
          maxLength: 64
          description: Current durable resource revision.
          example: rv-42
        changed:
          type: boolean
          description: True when the current revision is newer than the supplied checkpoint.
          example: true
    tenant-list:
      title: TenantList
      type: object
      description: One page from a stable tenant snapshot.
      additionalProperties: false
      required:
        - items
        - page
      properties:
        items:
          type: array
          example:
            - id: 11111111-1111-4111-8111-111111111111
              name: production
              labels: {}
              revision: 1
              createdAt: '2026-08-25T12:00:00Z'
              updatedAt: '2026-08-25T12:00:00Z'
          items:
            $ref: '#/components/schemas/tenant'
        page:
          $ref: '#/components/schemas/page-info'
...
  responses:
    unauthorized:
      description: Bearer authentication is missing or invalid.
      headers:
        WWW-Authenticate:
          $ref: '#/components/headers/www-authenticate'
        X-Request-ID:
          $ref: '#/components/headers/request-id'
      content:
        application/problem+json:
          example:
            type: https://moco.araihu.com/problems/unauthorized
            title: Unauthorized
            status: 401
            code: unauthorized
            requestId: 11111111-1111-4111-8111-111111111111
          schema:
            $ref: '#/components/schemas/problem'
    forbidden:
      description: The authenticated principal lacks permission for the operation.
      headers:
        X-Request-ID:
          $ref: '#/components/headers/request-id'
      content:
        application/problem+json:
          example:
            type: https://moco.araihu.com/problems/forbidden
            title: Forbidden
            status: 403
            code: forbidden
            requestId: 11111111-1111-4111-8111-111111111111
          schema:
            $ref: '#/components/schemas/problem'
    too-many-requests:
      description: |
        The caller exceeded an enforced rate limit. A deployment may enforce this in
        Mocó or at its API gateway; clients must honor Retry-After before retrying.
      headers:
        Retry-After:
          $ref: '#/components/headers/retry-after'
        X-Request-ID:
          $ref: '#/components/headers/request-id'
      content:
        application/problem+json:
          example:
            type: https://moco.araihu.com/problems/too-many-requests
            title: Too Many Requests
            status: 429
            code: rate_limited
            requestId: 11111111-1111-4111-8111-111111111111
          schema:
            $ref: '#/components/schemas/problem'
    internal-error:
      description: The server failed to complete the request.
      headers:
        X-Request-ID:
          $ref: '#/components/headers/request-id'
      content:
        application/problem+json:
          example:
            type: https://moco.araihu.com/problems/internal-error
            title: Internal Server Error
            status: 500
            code: internal_error
            requestId: 11111111-1111-4111-8111-111111111111
          schema:
            $ref: '#/components/schemas/problem'
    service-unavailable:
      description: |
        The service or a required dependency is temporarily unavailable. This may be
        emitted by Mocó or by the deployment edge; clients should retry after the
        advertised interval with bounded exponential backoff.
      headers:
        Retry-After:
          $ref: '#/components/headers/retry-after'
        X-Request-ID:
          $ref: '#/components/headers/request-id'
      content:
        application/problem+json:
          example:
            type: https://moco.araihu.com/problems/service-unavailable
            title: Service Unavailable
            status: 503
            code: service_unavailable
            requestId: 11111111-1111-4111-8111-111111111111
          schema:
            $ref: '#/components/schemas/problem'
    bad-request:
      description: Request parameters, headers, or body are invalid.
      headers:
        X-Request-ID:
          $ref: '#/components/headers/request-id'
      content:
        application/problem+json:
          example:
            type: https://moco.araihu.com/problems/bad-request
            title: Bad Request
            status: 400
            code: invalid_request
            requestId: 11111111-1111-4111-8111-111111111111
          schema:
            $ref: '#/components/schemas/problem'
...
  parameters:
    request-id:
      name: X-Request-ID
      in: header
      required: false
      description: Optional caller correlation ID; the server returns the effective ID.
      schema:
        type: string
        minLength: 1
        maxLength: 128
    resource-version:
      name: resourceVersion
      in: query
      required: false
      description: |
        Opaque durable change checkpoint returned by the watch endpoint. A caller
        that supplies a checkpoint is notified when any persisted resource changes;
        it should then resync the collections it is responsible for. Because the
        checkpoint is process-wide, callers need the cluster-wide watch permission.
      schema:
        type: string
        pattern: '^rv-[0-9]+$'
        minLength: 3
        maxLength: 64
        example: rv-42
    watch-timeout:
      name: timeoutSeconds
      in: query
      required: false
      description: Maximum number of seconds to wait for a change. Zero returns immediately. The limit leaves headroom for the server write timeout.
      schema:
        type: integer
        format: int32
        minimum: 0
        maximum: 25
        default: 0
        example: 15
    limit:
      name: limit
      in: query
      required: false
      description: Maximum number of items to return.
      schema:
        type: integer
        format: int32
        minimum: 1
        maximum: 200
        default: 50
    cursor:
      name: cursor
      in: query
      required: false
      description: Opaque next-page cursor returned by the preceding list response. Repeat the same `prefix` query value when continuing a path-scoped secret listing.
      schema:
        type: string
        minLength: 1
        maxLength: 2048
    name-filter:
      name: name
      in: query
      required: false
      description: Exact, case-sensitive resource name filter.
      schema:
        type: string
        minLength: 1
        maxLength: 128
    external-id-filter:
      name: externalId
      in: query
      required: false
      description: Exact immutable external correlation identifier filter.
      schema:
        type: string
        minLength: 1
        maxLength: 512
...
    tenant-id:
      name: tenantId
      in: path
      required: true
      description: Tenant UUID.
      schema:
        type: string
        format: uuid
...
  headers:
    request-id:
      description: Server-selected request correlation identifier.
      required: true
      schema:
        type: string
        minLength: 1
      example: 11111111-1111-4111-8111-111111111111
...
    cache-control:
      description: Cache directive for secret-bearing responses.
      required: true
      schema:
        type: string
        enum:
          - no-store
      example: no-store
...